+++ ui/src/kapacitor/constants/index.js
@@ -40,6 +40,7 @@ export const ALERTS = [
   'hipchat',
   'opsgenie',
   'pagerduty',
+  'pushover',
   'sensu',
   'slack',
   'smtp',
@@ -82,23 +83,64 @@ export const RULE_MESSAGE_TEMPLATES = {
 
 export const DEFAULT_ALERTS = ['http', 'tcp', 'exec', 'log']
 
-export const DEFAULT_ALERT_LABELS = {
-  http: 'URL:',
-  tcp: 'Address:',
-  exec: 'Add Command (Arguments separated by Spaces):',
-  log: 'File:',
-  smtp: 'Email Addresses (Separated by Spaces):',
-  slack: 'Send alerts to Slack channel:',
-  alerta: 'Paste Alerta TICKscript:',
-}
-export const DEFAULT_ALERT_PLACEHOLDERS = {
-  http: 'Ex: http://example.com/api/alert',
-  tcp: 'Ex: exampleendpoint.com:5678',
-  exec: 'Ex: woogie boogie',
-  log: 'Ex: /tmp/alerts.log',
-  smtp: 'Ex: benedict@domain.com delaney@domain.com susan@domain.com',
-  slack: '#alerts',
-  alerta: 'alerta()',
+export const RULE_ALERT_OPTIONS = {
+  http: {
+    args: {
+      label: 'URL:',
+      placeholder: 'Ex: http://example.com/api/alert',
+    },
+  },
+  tcp: {
+    args: {
+      label: 'Address:',
+      placeholder: 'Ex: exampleendpoint.com:5678',
+    },
+  },
+  exec: {
+    args: {
+      label: 'Add Command (Arguments separated by Spaces):',
+      placeholder: 'Ex: woogie boogie',
+    },
+  },
+  log: {
+    args: {
+      label: 'File:',
+      placeholder: 'Ex: /tmp/alerts.log',
+    },
+  },
+  smtp: {
+    args: {
+      label: 'Email Addresses (Separated by Spaces):',
+      placeholder:
+        'Ex: benedict@domain.com delaney@domain.com susan@domain.com',
+    },
+    details: {placeholder: 'Email body text goes here'},
+  },
+  slack: {
+    args: {
+      label: 'Send alerts to Slack channel:',
+      placeholder: '#alerts',
+    },
+  },
+  alerta: {
+    args: {
+      label: 'Paste Alerta TICKscript:',
+      placeholder: 'alerta()',
+    },
+  },
+  pushover: {
+    properties: [
+      {
+        name: 'device',
+        label: 'Device:',
+        placeholder: 'dv1,dv2 (Comma Separated)',
+      },
+      {name: 'title', label: 'Title:', placeholder: 'Important Message'},
+      {name: 'URL', label: 'URL:', placeholder: 'https://influxdata.com'},
+      {name: 'URLTitle', label: 'URL Title:', placeholder: 'InfluxData'},
+      {name: 'sound', label: 'Sound:', placeholder: 'alien'},
+    ],
+  },
 }
 
 export const ALERT_NODES_ACCESSORS = {